api/requirements.txt
fastapi[standard]
uvicorn
python-multipart
python-dotenv
pillow
albumentations
opencv-python-headless
numpy
requests
google-genai
